+++ lib/td_dd/data_structures/data_structures.ex
@@ -20,7 +20,6 @@ defmodule TdDd.DataStructures do
   alias TdDd.DataStructures.DataStructureRelation
   alias TdDd.DataStructures.DataStructuresTags
   alias TdDd.DataStructures.DataStructureTag
-  alias TdDd.DataStructures.DataStructureType
   alias TdDd.DataStructures.DataStructureVersion
   alias TdDd.DataStructures.StructureMetadata
   alias TdDd.DataStructures.StructureNote
@@ -34,7 +33,7 @@ defmodule TdDd.DataStructures do
   # Data structure version associations preloaded for some views
   @preload_dsv_assocs [:classifications, data_structure: :system]
 
-  def list_data_structures(clauses \\ %{}, options \\ []) do
+  def list_data_structures(clauses \\ %{}) do
     clauses
     |> Enum.reduce(DataStructure, fn
       {:external_id, external_ids}, q when is_list(external_ids) ->
@@ -58,7 +57,17 @@ defmodule TdDd.DataStructures do
     |> select_merge([_, sn], %{latest_note: sn.df_content})
     |> preload(:system)
     |> Repo.all()
-    |> Enum.map(&enrich(&1, options))
+  end
+
+  def list_data_structure_versions(clauses \\ %{}) do
+    clauses
+    |> Enum.reduce(DataStructureVersion, fn
+      {:since, since}, q -> where(q, [dsv], dsv.updated_at >= ^since or dsv.deleted_at >= ^since)
+      {:min_id, id}, q -> where(q, [dsv], dsv.id >= ^id)
+      {:order_by, "id"}, q -> order_by(q, :id)
+      {:limit, limit}, q -> limit(q, ^limit)
+    end)
+    |> Repo.all()
   end
 
   def get_data_structure!(id) do
@@ -84,18 +93,18 @@ defmodule TdDd.DataStructures do
     enriched_structure_version!(id)
   end
 
-  def get_data_structure_version!(data_structure_version_id, options) do
+  def get_data_structure_version!(data_structure_version_id, opts) do
     data_structure_version_id
     |> get_data_structure_version!()
-    |> enrich(options)
+    |> enrich(opts)
   end
 
-  def get_data_structure_version!(data_structure_id, version, options) do
+  def get_data_structure_version!(data_structure_id, version, opts) do
     DataStructureVersion
     |> Repo.get_by!(data_structure_id: data_structure_id, version: version)
     |> Map.get(:id)
     |> enriched_structure_version!(preload: @preload_dsv_assocs)
-    |> enrich(options)
+    |> enrich(opts)
   end
 
   def get_cached_content(%{} = content, %{data_structure_type: %{template_id: template_id}}) do
@@ -114,75 +123,66 @@ defmodule TdDd.DataStructures do
 
   defp enrich(target, nil = _opts), do: target
 
-  defp enrich(%DataStructure{} = ds, options) do
-    ds
-    |> StructureEnricher.enrich()
-    |> enrich(options, :versions, &Repo.preload(&1, :versions))
-    |> enrich(options, :latest, &get_latest_version/1)
-    |> enrich(options, :source, &get_source/1)
-    |> enrich(options, :metadata_versions, &get_metadata_versions/1)
-  end
-
   defp enrich(%DataStructureVersion{id: id} = _data_structure_version, :defaults) do
     enriched_structure_version!(id, preload: [data_structure: :source])
   end
 
-  defp enrich(%DataStructureVersion{} = dsv, options) do
+  defp enrich(%DataStructureVersion{} = dsv, opts) do
     deleted = not is_nil(Map.get(dsv, :deleted_at))
-    with_confidential = Enum.member?(options, :with_confidential)
+    with_confidential = Enum.member?(opts, :with_confidential)
 
     dsv
     |> enrich(:defaults)
-    |> enrich(options, :classifications, &get_classifications/1)
-    |> enrich(options, :system, &get_system/1)
+    |> enrich(opts, :classifications, &get_classifications!/1)
+    |> enrich(opts, :system, &get_system!/1)
     |> enrich(
-      options,
+      opts,
       :parents,
       &get_parents(&1, deleted: deleted, with_confidential: with_confidential)
     )
     |> enrich(
-      options,
+      opts,
       :children,
       &get_children(&1, deleted: deleted, with_confidential: with_confidential)
     )
     |> enrich(
-      options,
+      opts,
       :siblings,
       &get_siblings(&1, deleted: deleted, with_confidential: with_confidential)
     )
     |> enrich(
-      options,
+      opts,
       :data_fields,
       &get_field_structures(&1,
         deleted: deleted,
-        preload: if(Enum.member?(options, :profile), do: [data_structure: :profile], else: []),
+        preload: if(Enum.member?(opts, :profile), do: [data_structure: :profile], else: []),
         with_confidential: with_confidential
       )
     )
-    |> enrich(options, :data_field_degree, &get_field_degree/1)
-    |> enrich(options, :data_field_links, &get_field_links/1)
+    |> enrich(opts, :data_field_degree, &get_field_degree/1)
+    |> enrich(opts, :data_field_links, &get_field_links/1)
     |> enrich(
-      options,
+      opts,
       :relations,
       &get_relations(&1, deleted: deleted, default: false, with_confidential: with_confidential)
     )
-    |> enrich(options, :relation_links, &get_relation_links/1)
-    |> enrich(options, :versions, &get_versions/1)
-    |> enrich(options, :degree, &get_degree/1)
-    |> enrich(options, :profile, &get_profile/1)
-    |> enrich(options, :links, &get_structure_links/1)
-    |> enrich(options, :source, &get_source/1)
-    |> enrich(options, :metadata_versions, &get_metadata_versions/1)
-    |> enrich(options, :data_structure_type, &get_data_structure_type/1)
-    |> enrich(options, :tags, &get_tags/1)
-    |> enrich(options, :grants, &get_grants/1)
-    |> enrich(options, :grant, &get_grant(&1, options[:user_id]))
+    |> enrich(opts, :relation_links, &get_relation_links/1)
+    |> enrich(opts, :versions, &get_versions!/1)
+    |> enrich(opts, :degree, &get_degree/1)
+    |> enrich(opts, :profile, &get_profile!/1)
+    |> enrich(opts, :links, &get_structure_links/1)
+    |> enrich(opts, :source, &get_source!/1)
+    |> enrich(opts, :metadata_versions, &get_metadata_versions!/1)
+    |> enrich(opts, :data_structure_type, &get_data_structure_type!/1)
+    |> enrich(opts, :tags, &get_tags!/1)
+    |> enrich(opts, :grants, &get_grants/1)
+    |> enrich(opts, :grant, &get_grant(&1, opts[:user_id]))
   end
 
-  defp enrich(%{} = target, options, key, fun) do
+  defp enrich(%{} = target, opts, key, fun) do
     target_key = get_target_key(key)
 
-    case Enum.member?(options, key) do
+    case Enum.member?(opts, key) do
       false -> target
       true -> Map.put(target, target_key, fun.(target))
     end
@@ -193,63 +193,61 @@ defmodule TdDd.DataStructures do
   defp get_target_key(:relation_links), do: :relations
   defp get_target_key(key), do: key
 
-  defp get_system(%DataStructureVersion{} = dsv) do
-    dsv
-    |> Repo.preload(data_structure: :system)
-    |> Map.get(:data_structure)
-    |> Map.get(:system)
+  defp get_system!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, data_structure: :system) do
+      %{data_structure: %{system: system}} -> system
+    end
   end
 
-  defp get_classifications(%DataStructureVersion{} = dsv) do
-    dsv
-    |> Repo.preload(:classifications)
-    |> Map.get(:classifications)
+  defp get_classifications!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, :classifications) do
+      %{classifications: classifications} -> classifications
+    end
   end
 
-  defp get_profile(%DataStructureVersion{} = dsv) do
-    dsv
-    |> Repo.preload(data_structure: :profile)
-    |> Map.get(:data_structure)
-    |> Map.get(:profile)
+  defp get_profile!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, data_structure: :profile) do
+      %{data_structure: %{profile: profile}} -> profile
+    end
   end
 
-  def get_data_structure_type(%DataStructureVersion{} = dsv) do
-    DataStructureType
-    |> where([ds_type], ds_type.name == ^dsv.type)
-    |> Repo.one()
+  def get_data_structure_type!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, :structure_type) do
+      %{structure_type: structure_type} -> structure_type
+    end
   end
 
-  def get_field_structures(data_structure_version, options) do
+  def get_field_structures(data_structure_version, opts) do
     data_structure_version
     |> Ecto.assoc(:children)
     |> where([child], child.class == "field")
     |> join(:inner, [child], child in assoc(child, :data_structure))
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic([_child, _parent, child_ds], child_ds.confidential == false)
     )
-    |> with_deleted(options, dynamic([child], is_nil(child.deleted_at)))
+    |> with_deleted(opts, dynamic([child], is_nil(child.deleted_at)))
     |> select([child], child)
     |> Repo.all()
-    |> Repo.preload(options[:preload] || [])
+    |> Repo.preload(opts[:preload] || [])
   end
 
-  def get_children(%DataStructureVersion{id: id}, options \\ []) do
+  def get_children(%DataStructureVersion{id: id}, opts \\ []) do
     DataStructureRelation
     |> where([r], r.parent_id == ^id)
     |> join(:inner, [r], parent in assoc(r, :child))
     |> join(:inner, [r, _parent, _], r in assoc(r, :relation_type))
     |> join(:inner, [_r, child, _], child in assoc(child, :data_structure))
     |> with_deleted(
-      Keyword.get(options, :deleted),
+      Keyword.get(opts, :deleted),
       dynamic([_, parent, _], is_nil(parent.deleted_at))
     )
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic([_, _child, _, ds_child], ds_child.confidential == false)
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic([_, _child, relation_type, _], relation_type.name == "default"),
       dynamic([_, _child, relation_type, _], relation_type.name != "default")
     )
@@ -261,26 +259,26 @@ defmodule TdDd.DataStructures do
       relation_type: relation_type
     })
     |> Repo.all()
-    |> select_structures(Keyword.get(options, :default))
+    |> select_structures(Keyword.get(opts, :default))
   end
 
-  def get_parents(%DataStructureVersion{id: id}, options \\ []) do
+  def get_parents(%DataStructureVersion{id: id}, opts \\ []) do
     DataStructureRelation
     |> where([r], r.child_id == ^id)
     |> join(:inner, [r], parent in assoc(r, :parent))
     |> join(:inner, [r, _parent], relation_type in assoc(r, :relation_type))
     |> join(:inner, [_r, parent, _relation_type], parent in assoc(parent, :data_structure))
     |> with_deleted(
-      Keyword.get(options, :deleted),
+      Keyword.get(opts, :deleted),
       dynamic([_, parent, _, _], is_nil(parent.deleted_at))
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic([_, _parent, relation_type, _], relation_type.name == "default"),
       dynamic([_, _parent, relation_type, _], relation_type.name != "default")
     )
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic([_, _parent, _relation_type, parent_ds], parent_ds.confidential == false)
     )
     |> order_by([_, parent, _, _], asc: parent.data_structure_id, desc: parent.version)
@@ -291,10 +289,10 @@ defmodule TdDd.DataStructures do
       relation_type: relation_type
     })
     |> Repo.all()
-    |> select_structures(Keyword.get(options, :default))
+    |> select_structures(Keyword.get(opts, :default))
   end
 
-  def get_siblings(%DataStructureVersion{id: id}, options \\ []) do
+  def get_siblings(%DataStructureVersion{id: id}, opts \\ []) do
     DataStructureRelation
     |> where([r], r.child_id == ^id)
     |> join(:inner, [r], parent in assoc(r, :parent))
@@ -310,28 +308,28 @@ defmodule TdDd.DataStructures do
       sibling in assoc(sibling, :data_structure)
     )
     |> with_deleted(
-      options,
+      opts,
       dynamic(
         [_r, parent, _parent_rt, _r_c, _child_rt, _sibling, _sibling_ds],
         is_nil(parent.deleted_at)
       )
     )
     |> with_deleted(
-      options,
+      opts,
       dynamic(
         [_r, parent, _parent_rt, _r_c, _child_rt, sibling, _sibling_ds],
         is_nil(sibling.deleted_at)
       )
     )
     |> with_confidential(
-      Keyword.get(options, :with_confidential),
+      Keyword.get(opts, :with_confidential),
       dynamic(
         [_r, _parent, _parent_rt, _r_c, _child_rt, _sibling, sibling_ds],
         sibling_ds.confidential == false
       )
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic(
         [_r, _parent, parent_rt, _r_c, _child_rt, _sibling, _sibling_ds],
         parent_rt.name == "default"
@@ -342,7 +340,7 @@ defmodule TdDd.DataStructures do
       )
     )
     |> relation_type_condition(
-      Keyword.get(options, :default),
+      Keyword.get(opts, :default),
       dynamic(
         [_r, _parent, _parent_rt, _r_c, child_rt, _sibling, _sibling_ds],
         child_rt.name == "default"
@@ -363,9 +361,9 @@ defmodule TdDd.DataStructures do
     |> Enum.uniq_by(& &1.data_structure_id)
   end
 
-  defp get_relations(%DataStructureVersion{} = version, options) do
-    parents = get_parents(version, options)
-    children = get_children(version, options)
+  defp get_relations(%DataStructureVersion{} = version, opts) do
+    parents = get_parents(version, opts)
+    children = get_children(version, opts)
 
     %{parents: parents, children: children}
   end
@@ -388,10 +386,10 @@ defmodule TdDd.DataStructures do
     |> Map.put(:parents, parents)
   end
 
-  defp get_versions(%DataStructureVersion{} = dsv) do
-    dsv
-    |> Ecto.assoc([:data_structure, :versions])
-    |> Repo.all()
+  defp get_versions!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, data_structure: :versions) do
+      %{data_structure: %{versions: versions}} -> versions
+    end
   end
 
   defp get_grants(%DataStructureVersion{data_structure_id: id, path: path}, clauses \\ %{}) do
@@ -437,52 +435,24 @@ defmodule TdDd.DataStructures do
     Map.put(version, :version, Repo.preload(v, preloads))
   end
 
-  def get_source(%DataStructureVersion{data_structure: %DataStructure{} = data_structure}) do
-    get_source(data_structure)
+  defp get_source!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, data_structure: :source) do
+      %{data_structure: %{source: source}} -> source
+    end
   end
 
-  def get_source(%DataStructureVersion{} = version) do
-    version
-    |> Repo.preload(data_structure: :source)
-    |> Kernel.get_in([:data_structure, :source])
+  defp get_metadata_versions!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv, data_structure: :metadata_versions) do
+      %{data_structure: %{metadata_versions: metadata_versions}} -> metadata_versions
+    end
   end
 
-  def get_source(%DataStructure{source: %NotLoaded{}} = data_structure) do
-    data_structure
-    |> Repo.preload(:source)
-    |> Map.get(:source)
-  end
-
-  def get_source(%DataStructure{source: source}), do: source
-
-  defp get_metadata_versions(%DataStructure{} = data_structure) do
-    data_structure
-    |> Repo.preload(:metadata_versions)
-    |> Repo.get(:metadata_versions)
-  end
-
-  defp get_metadata_versions(%DataStructureVersion{} = version) do
-    version
-    |> Repo.preload(data_structure: :metadata_versions)
-    |> Map.get(:data_structure)
-    |> Map.get(:metadata_versions)
-  end
-
-  defp get_tags(%DataStructureVersion{data_structure: %DataStructure{} = data_structure}) do
-    get_tags(data_structure)
-  end
-
-  defp get_tags(%DataStructureVersion{} = version) do
-    version
-    |> Repo.preload(:data_structure)
-    |> Map.get(:data_structure)
-    |> get_tags()
-  end
-
-  defp get_tags(%DataStructure{} = data_structure) do
-    data_structure
-    |> Repo.preload(data_structures_tags: [:data_structure_tag, :data_structure])
-    |> Map.get(:data_structures_tags)
+  defp get_tags!(%DataStructureVersion{} = dsv) do
+    case Repo.preload(dsv,
+           data_structure: [data_structures_tags: [:data_structure_tag, :data_structure]]
+         ) do
+      %{data_structure: %{data_structures_tags: data_structures_tags}} -> data_structures_tags
+    end
   end
 
   @doc """
@@ -636,22 +606,22 @@ defmodule TdDd.DataStructures do
      }}
   end
 
-  def get_latest_version(target, options \\ [])
+  def get_latest_version(target, opts \\ [])
 
   def get_latest_version(nil, _), do: nil
 
-  def get_latest_version(%DataStructure{id: id}, options) do
-    get_latest_version(id, options)
+  def get_latest_version(%DataStructure{id: id}, opts) do
+    get_latest_version(id, opts)
   end
 
-  def get_latest_version(data_structure_id, options) do
+  def get_latest_version(data_structure_id, opts) do
     DataStructureVersion
     |> where(data_structure_id: ^data_structure_id)
     |> distinct(:data_structure_id)
     |> order_by(desc: :version)
     |> preload(data_structure: :source)
     |> Repo.one()
-    |> enrich(options)
+    |> enrich(opts)
   end
 
   def put_domain_id(data, %{} = domain_map, domain) when is_binary(domain) do
@@ -764,16 +734,16 @@ defmodule TdDd.DataStructures do
     end
   end
 
-  def get_latest_version_by_external_id(external_id, options \\ []) do
+  def get_latest_version_by_external_id(external_id, opts \\ []) do
     DataStructureVersion
-    |> with_deleted(options, dynamic([dsv], is_nil(dsv.deleted_at)))
+    |> with_deleted(opts, dynamic([dsv], is_nil(dsv.deleted_at)))
     |> distinct(:data_structure_id)
     |> order_by(desc: :version)
     |> join(:inner, [data_structure], ds in assoc(data_structure, :data_structure))
     |> where([_, ds], ds.external_id == ^external_id)
     |> select_merge([_, ds], %{external_id: ds.external_id})
     |> Repo.one()
-    |> enrich(options[:enrich])
+    |> enrich(opts[:enrich])
   end
 
   @spec get_latest_versions([non_neg_integer]) :: [DataStructureVersion.t()]
@@ -785,8 +755,8 @@ defmodule TdDd.DataStructures do
     |> Repo.all()
   end
 
-  defp with_deleted(query, options, dynamic) when is_list(options) do
-    include_deleted = Keyword.get(options, :deleted, true)
+  defp with_deleted(query, opts, dynamic) when is_list(opts) do
+    include_deleted = Keyword.get(opts, :deleted, true)
     with_deleted(query, include_deleted, dynamic)
   end
 
@@ -923,16 +893,16 @@ defmodule TdDd.DataStructures do
 
   defp do_profile_source(dsv, _source), do: dsv
 
-  def list_data_structure_tags(options \\ []) do
+  def list_data_structure_tags(opts \\ []) do
     DataStructureTag
     |> Repo.all()
-    |> Repo.preload(options[:preload] || [])
+    |> Repo.preload(opts[:preload] || [])
   end
 
-  def get_data_structure_tag!(id, options \\ []) do
+  def get_data_structure_tag!(id, opts \\ []) do
     DataStructureTag
     |> Repo.get!(id)
-    |> Repo.preload(options[:preload] || [])
+    |> Repo.preload(opts[:preload] || [])
   end
 
   def create_data_structure_tag(attrs \\ %{}) do
@@ -1208,7 +1178,12 @@ defmodule TdDd.DataStructures do
 
   """
   def create_structure_note(data_structure, attrs, user_id) do
-    changeset = StructureNote.create_changeset(%StructureNote{}, data_structure, attrs)
+    changeset =
+      StructureNote.create_changeset(
+        %StructureNote{},
+        data_structure,
+        attrs
+      )
 
     Multi.new()
     |> Multi.insert(:structure_note, changeset)
@@ -1267,6 +1242,45 @@ defmodule TdDd.DataStructures do
 
   """
 
+  def update_structure_note(
+        %StructureNote{} = structure_note,
+        %{"status" => status} = attrs,
+        user_id
+      )
+      when status in [
+             "published",
+             "pending_approval",
+             "rejected",
+             "published",
+             "versioned",
+             "draft",
+             "deprecated"
+           ] do
+    changeset = StructureNote.changeset(structure_note, attrs)
+
+    %{data_structure: data_structure} =
+      structure_note =
+      structure_note
+      |> Repo.preload(:data_structure)
+
+    Multi.new()
+    |> Multi.run(:latest, fn _, _ ->
+      {:ok, get_latest_version(data_structure, [:path])}
+    end)
+    |> Multi.run(:structure_note, fn _, _ ->
+      {:ok, structure_note}
+    end)
+    |> Multi.update(:structure_note_update, changeset)
+    |> Multi.run(:audit, Audit, :structure_note_status_updated, [status, user_id])
+    |> Repo.transaction()
+    |> case do
+      {:ok, res} -> {:ok, Map.get(res, :structure_note_update)}
+      {:error, :structure_note_update, err, _} -> {:error, err}
+      err -> err
+    end
+    |> on_update()
+  end
+
   def update_structure_note(%StructureNote{} = structure_note, attrs, user_id) do
     changeset = StructureNote.changeset(structure_note, attrs)
 
@@ -1287,15 +1301,36 @@ defmodule TdDd.DataStructures do
 
   ## Examples
 
-      iex> delete_structure_note(structure_note)
+      iex> delete_structure_note(structure_note, user_id)
       {:ok, %StructureNote{}}
 
-      iex> delete_structure_note(structure_note)
+      iex> delete_structure_note(structure_note, user_id)
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_structure_note(%StructureNote{} = structure_note) do
-    Repo.delete(structure_note)
+  def delete_structure_note(
+        %StructureNote{} = structure_note,
+        user_id
+      ) do
+    %{data_structure: data_structure} =
+      structure_note =
+      structure_note
+      |> Repo.preload(:data_structure)
+
+    Multi.new()
+    |> Multi.run(:latest, fn _, _ ->
+      {:ok, get_latest_version(data_structure, [:path])}
+    end)
+    |> Multi.delete(:structure_note, structure_note)
+    |> Multi.run(:audit, Audit, :structure_note_deleted, [user_id])
+    |> Repo.transaction()
+    |> case do
+      {:ok, %{structure_note: structure_note}} ->
+        {:ok, structure_note}
+
+      {:error, _, changeset, _} ->
+        {:error, changeset}
+    end
   end
 
   @doc """
@@ -1309,5 +1344,15 @@ defmodule TdDd.DataStructures do
   """
   def change_structure_note(%StructureNote{} = structure_note, attrs \\ %{}) do
     StructureNote.changeset(structure_note, attrs)
+  end
+
+  ## Dataloader
+
+  def datasource do
+    Dataloader.Ecto.new(TdDd.Repo, query: &query/2, timeout: Dataloader.default_timeout())
+  end
+
+  defp query(queryable, _params) do
+    queryable
   end
 end