+++ test/support/factory.ex
@@ -80,7 +80,7 @@ defmodule TdDd.Factory do
 
   def rule_factory do
     %TdDq.Rules.Rule{
-      business_concept_id: sequence(:business_concept_id, &"#{&1}"),
+      business_concept_id: sequence(:business_concept_id, & &1),
       domain_id: sequence(:domain_id, &"#{&1}"),
       description: %{"document" => "Rule Description"},
       goal: 30,
@@ -313,7 +313,8 @@ defmodule TdDd.Factory do
     %TdDq.Implementations.ConditionRow{
       value: [%{"raw" => 8}],
       operator: build(:operator),
-      structure: build(:dataset_structure)
+      structure: build(:dataset_structure),
+      population: []
     }
   end