+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [4.27.0] 2021-09-07
+
+### Added
+
+- [TD-3951] Include additional information for grant events
+- [TD-3484] GraphQL API on `/api/v2`
+- [TD-3972] Nested population in validations
+- [TD-3910] Notifications included for structures notes status changes
+- [TD-3546] `with_profiling` filter in data structure version
+- [TD-3983] renders request_grant permission on structure_version
+
+### Changed
+
+- [TD-3826] Data quality permissions now uses `domain_id` instead of
+  `business_concept_id`
+- [TD-3039] `GET /api/data_structures/:data_structure_id/versions/:id` now
+  includes mutable metadata in the `metadata` field. The `metadata_versions`
+  field is no longer included in the response. The `metadata_fields` field in
+  the data structure type responses now include mutable metadata fields.
+- [TD-3973] Update td-df-lib for default values in swith fields
+
 ## [4.26.0] 2021-08-16
 
 ### Added